fix: 修复编译错误
- 修复 JSONObject.put() 参数问题：拆分为独立语句
- 修复 connect() 参数类型：使用 (int) 强制转换
- 移除错误的 "default" 注解字段（JSON Schema 不支持该语法）

+++ app/src/main/java/com/stupidbeauty/sisterfuture/tool/RemoteCommandTool.java
@@ -17,7 +17,7 @@ import java.util.concurrent.TimeUnit;
 
 public class RemoteCommandTool implements Tool {
     private static final String TAG = "RemoteCommandTool";
-    private static final long DEFAULT_TIMEOUT_MS = 60000; // 60秒默认超时
+    private static final int DEFAULT_TIMEOUT_MS = 60000; // 60 秒默认超时
     
     private final Context context;
     private static final ExecutorService executor = Executors.newSingleThreadExecutor();
@@ -140,7 +140,7 @@ public class RemoteCommandTool implements Tool {
             }
             
             session.setConfig("StrictHostKeyChecking", "no");
-            session.connect(3000); // 3秒连接超时
+            session.connect(3000); // 3 秒连接超时
 
             channel = (ChannelExec) session.openChannel("exec");
             channel.setCommand(command);
